Feat : 주문생성, 주문조회 API 추가
 - build.gradle 수정 ( querydsl )
 - 요구사항 기본 기능 구현
 - schema.sql, OrderRunner 삭제 ( ddl 자동 생성 및 시점관련 삭제 )

+++ src/main/java/antigravity/model/request/ProductInfoRequest.java
@@ -7,8 +7,8 @@ import lombok.Data;
 @Builder
 public class ProductInfoRequest {
 
-    private int productId;
-    private int productCount;
-    private String color;
-    private String size;
+	private long productId;
+	private int productCount;
+	private String color;
+	private String size;
 }
